optmization helpers tcp_forward argument parser

+++ sevent/helpers/tcp_forward.py
@@ -288,11 +288,11 @@ if __name__ == '__main__':
     parser.add_argument('-L', dest='forwards', default=[], action="append", type=str,
                         help='forward host, accept format [[local_bind:]local_port:remote_host:remote_port], support muiti forward args (example: 0.0.0.0:80:127.0.0.1:8088)')
     parser.add_argument('-t', dest='timeout', default=7200, type=int, help='timeout (default: 7200)')
-    parser.add_argument('-s', dest='speed', default=0, type=lambda v: (int(v[:-1]) * BYTES_MAP[v.upper()[-1]]) \
-        if v and v.upper()[-1] in BYTES_MAP else int(v), help='per connection speed limit byte, example: 1024, 1M (default: 0 is unlimit)')
-    parser.add_argument('-S', dest='global_speed', default=0, type=lambda v: (int(v[:-1]) * BYTES_MAP[v.upper()[-1]]) \
-        if v and v.upper()[-1] in BYTES_MAP else int(v), help='global speed limit byte, example: 1024, 1M (default: 0 is unlimit)')
-    parser.add_argument('-d', dest='delay', default=0, type=int, help='delay millisecond (default: 0 is not delay)')
+    parser.add_argument('-s', dest='speed', default=0, type=lambda v: int(float(v[:-1]) * BYTES_MAP[v.upper()[-1]]) \
+        if v and v.upper()[-1] in BYTES_MAP else int(float(v)), help='per connection speed limit byte, example: 1024, 1M (default: 0 is unlimit)')
+    parser.add_argument('-S', dest='global_speed', default=0, type=lambda v: int(float(v[:-1]) * BYTES_MAP[v.upper()[-1]]) \
+        if v and v.upper()[-1] in BYTES_MAP else int(float(v)), help='global speed limit byte, example: 1024, 1M (default: 0 is unlimit)')
+    parser.add_argument('-d', dest='delay', default=0, type=float, help='delay millisecond (default: 0 is not delay)')
     args = parser.parse_args()
 
     if not args.forwards:
